Yearmonth field type deprecated in favour of just Date and Datetime, with DRF Datefield used to allow multi date input_types
