+++ onyx/data/filters.py
@@ -99,27 +99,12 @@ class NumberRangeFilter(BaseRangeFilter, filters.NumberFilter):
     pass
 
 
-class YearMonthForm(forms.DateField):
-    def __init__(self, **kwargs):
-        kwargs["input_formats"] = ["%Y-%m"]
-        super().__init__(**kwargs)
-
-
-class YearMonthFilter(filters.Filter):
-    field_class = YearMonthForm
-
-
-class YearMonthInFilter(filters.BaseInFilter, YearMonthFilter):
-    pass
-
-
-class YearMonthRangeFilter(BaseRangeFilter, YearMonthFilter):
-    pass
-
-
 class DateFieldForm(forms.DateField):
     def __init__(self, **kwargs):
-        kwargs["input_formats"] = ["%Y-%m-%d"]
+        kwargs["input_formats"] = [
+            "%Y-%m",
+            "%Y-%m-%d",
+        ]
         super().__init__(**kwargs)
 
     def clean(self, value):
@@ -144,6 +129,7 @@ class DateRangeFilter(BaseRangeFilter, DateFilter):
 class DateTimeFieldForm(forms.DateTimeField):
     def __init__(self, **kwargs):
         kwargs["input_formats"] = [
+            "%Y-%m",
             "%Y-%m-%d",
             "%Y-%m-%d %H:%M",
             "%Y-%m-%d %H:%M:%S",
@@ -242,32 +228,10 @@ FILTERS = {
         "range": NumberRangeFilter,
         "isnull": IsNullFilter,
     },
-    OnyxType.DATE_YYYY_MM: {
-        lookup: YearMonthFilter for lookup in OnyxType.DATE_YYYY_MM.lookups
-    }
-    | {
-        "in": YearMonthInFilter,
-        "range": YearMonthRangeFilter,
-        "year": filters.NumberFilter,
-        "year__in": NumberInFilter,
-        "year__range": NumberRangeFilter,
-        "iso_year": filters.NumberFilter,
-        "iso_year__in": NumberInFilter,
-        "iso_year__range": NumberRangeFilter,
-        "week": filters.NumberFilter,
-        "week__in": NumberInFilter,
-        "week__range": NumberRangeFilter,
-        "isnull": IsNullFilter,
-    },
-    OnyxType.DATE_YYYY_MM_DD: {
-        lookup: DateFilter for lookup in OnyxType.DATE_YYYY_MM_DD.lookups
-    }
+    OnyxType.DATE: {lookup: DateFilter for lookup in OnyxType.DATE.lookups}
     | {
         "in": DateInFilter,
         "range": DateRangeFilter,
-        "year": filters.NumberFilter,
-        "year__in": NumberInFilter,
-        "year__range": NumberRangeFilter,
         "iso_year": filters.NumberFilter,
         "iso_year__in": NumberInFilter,
         "iso_year__range": NumberRangeFilter,
@@ -280,9 +244,6 @@ FILTERS = {
     | {
         "in": DateTimeInFilter,
         "range": DateTimeRangeFilter,
-        "year": filters.NumberFilter,
-        "year__in": NumberInFilter,
-        "year__range": NumberRangeFilter,
         "iso_year": filters.NumberFilter,
         "iso_year__in": NumberInFilter,
         "iso_year__range": NumberRangeFilter,
